+++ src/code_searcher/code_searcher_class.py
@@ -4,7 +4,7 @@ code_searcher.code_searcher_class
 ~~~~~~~~~~~~
 
 This module is consists of main class for code searching through your
-library
+code library
 
 :copyright: © 2019 by Stanislav Prokopyev stas.prokopiev@gmail.com.
 :license: MIT, see LICENSE.rst for more details.
@@ -15,15 +15,27 @@ import os
 
 from collections import defaultdict
 from collections import OrderedDict
-from collections import Counter
 
-
+#####
+# .decorators
 from code_searcher.decorators import check_type_of_arguments
-from code_searcher.working_with_files import get_dict_str_full_file_by_rel_path
-from code_searcher.additional_functions import get_names_of_all_functions_defined_in_py_code
-from code_searcher.additional_functions import get_list_modules_imported_in_py_code
+#####
+# .additional_functions
+from code_searcher.additional_functions import \
+    get_names_of_all_functions_defined_in_py_code
+from code_searcher.additional_functions import \
+    get_number_of_lines_in_string
+from code_searcher.additional_functions import \
+    get_list_modules_imported_in_py_code
+#####
+# .working_with_files
 from code_searcher.working_with_files import \
     get_list_str_filenames_of_all_files_with_given_extension
+from code_searcher.working_with_files import \
+    get_list_str_path_all_files_with_given_extension
+from code_searcher.working_with_files import \
+    get_file_as_string
+#####
 
 
 class code_searcher_class():
@@ -34,39 +46,60 @@ class code_searcher_class():
 
     Attributes
     ----------
-    list_str_folders_where_to_look : list
+    list_str_dirs_where_to_look : list
         list most parent folders where to look for code files
     list_str_file_extensions : list
         list strings file extensions which files to explore
-    dict_dict_str_file_by_filename_by_folder : dict
-        {folder_path: {relative_file_path: string_whole_file, ...}, ...}
+    dict_time_file_changed_by_path : dict
+        {"path_to_file_1": float_time_when_last_modified, ...}
+    dict_list_file_paths_by_ext_by_dir : dict
+        dict with pathes of all initialized files
+        {"folder_path_1": {"file_extension_1": [file_path_1, ...], ...}, ...}
+    dict_str_file_by_path_by_ext_by_dir : dict
+        {"dir_path_1": {"file_ext_1": {"file_path_1": whole_file, ..}, ..}, ..}
 
     Methods
     -------
+    get_file_stats_of_the_code_library(bool_is_to_print=False)
+        Getting string with statistics about files in the library
+
+    get_dict_list_file_paths_by_ext_by_dir()
+        Getting dict with pathes to asked files by ext by dir
+
+    download_files()
+        For all files defined in self.dict_list_file_paths_by_ext_by_dir
+        download up to date versions of files (Efficient realization)
+
+    update_files()
+        Re-search for all files in the folders and download them
+
     search_all_occurrences_of_the_code_in_the_library(
     str_code_to_search,
     bool_is_to_search_case_sensitive=True,
     )
         Searching some code inside whole library and
-        print all occurences in formatted way
+        print all occurences in nice formatted way
 
-    update_files()
-        Redownloading all files if something was changed inside them
-
-    get_file_stats_of_the_code_library(bool_is_to_print=False)
-        Getting string statistic about files in the library
-
-    get_dict_times_functions_used()
+    get_dict_times_py_functions_used()
         Getting {function_name: times_function_used, ...}
+
+    get_names_of_all_py_functions_defined_but_never_used()
+        Getting list of functions defined inside code but never used
 
     get_number_of_lines_in_the_library()
         Getting number of not empty lines in whole library
 
-    get_dict_num_of_code_lines_by_extension()
-        Getting dict number of not empty lines by file extension type
+    get_names_of_all_functions_defined_in_the_py_library
+        Getting set names of all functions defined in the library
 
-    get_names_of_all_functions_defined_but_never_used()
-        Getting list of functions defined inside code but never used
+    get_set_str_names_of_all_py_files
+        Getting set names of all .py files inside the library
+
+    get_list_all_imported_py_modules_in_the_library
+        Getting list of all modules imported in the library
+
+
+
 
     get_list_of_all_outer_modules_used_in_the_library()
         Getting list of all OUTER modules imported in the library
@@ -77,29 +110,47 @@ class code_searcher_class():
     @check_type_of_arguments
     def __init__(
             self,
-            list_str_folders_where_to_look,
+            list_str_dirs_where_to_look,
             list_str_file_extensions=[".py", "ipynb"],
     ):
         """
         Parameters
         ----------
-        list_str_folders_where_to_look : list
+        list_str_dirs_where_to_look : list
             list most parent folders where to look for code files
         list_str_file_extensions : list, optional
             list strings file extensions which files to explore
             (default is [".py", "ipynb"])
         """
         # 0) Check that as least some arguments are given
-        assert list_str_folders_where_to_look is not None, (
+        assert list_str_dirs_where_to_look is not None, (
             "ERROR: for initializing code_searcher obj should be given \n"
             "list of most parent folders where to search files with code \n"
-            "var list_str_folders_where_to_look: SHOULD NOT BE NONE"
+            "var list_str_dirs_where_to_look: SHOULD NOT BE NONE"
         )
         # 1) Initialize class variables
-        self.list_str_folders_where_to_look = list_str_folders_where_to_look
+        # list most parent folders where to look for code files
+        self.list_str_dirs_where_to_look = list_str_dirs_where_to_look
+        # list strings file extensions which files to explore
         self.list_str_file_extensions = list_str_file_extensions
-        # Initialize dict_dict_str_file_by_filename_by_folder
-        self.download_files()
+        # {"path_to_file_1": float_time_when_last_modified, ...}
+        self.dict_time_file_changed_by_path = defaultdict(float)
+        print("Downloading all files (it can be a long process, please wait.)")
+        float_time_start = time.time()
+        # {"folder_path_1": {"file_extension_1": [file_path_1, ...], ...}, ...}
+        self.dict_list_file_paths_by_ext_by_dir = \
+            self.get_dict_list_file_paths_by_ext_by_dir()
+        # {"dir_path_1": {"file_ext_1": {"file_path_1": whole_file, ..}, .}, .}
+        # Initialize dict_str_file_by_path_by_ext_by_dir
+        self.dict_str_file_by_path_by_ext_by_dir = OrderedDict()
+        int_files_downloaded = self.download_files()
+        float_seconds_spent = round(time.time() - float_time_start, 2)
+        print(
+            "{} Files were downloaded in {} seconds".format(
+                int_files_downloaded,
+                float_seconds_spent
+            )
+        )
         # # Print file statistic
         # self.get_file_stats_of_the_code_library(bool_is_to_print=True)
         # print("Search obj INITIALIZED.")
@@ -122,8 +173,8 @@ class code_searcher_class():
             "\n\n"
         )
         str_obj_repr += "Folders to search in: \n"
-        for str_folder in self.list_str_folders_where_to_look:
-            str_obj_repr += "--> " + str_folder + "\n"
+        for str_dir in self.list_str_dirs_where_to_look:
+            str_obj_repr += "--> " + str_dir + "\n"
         str_obj_repr += "Extensions to check: \n"
         for str_ext in self.list_str_file_extensions:
             str_obj_repr += "--> " + str_ext + "\n"
@@ -132,7 +183,7 @@ class code_searcher_class():
 
     @check_type_of_arguments
     def get_file_stats_of_the_code_library(self, bool_is_to_print=False):
-        """Getting string statistic about files in the library
+        """Getting string with statistics about files in the library
 
         Parameters
         ----------
@@ -145,43 +196,47 @@ class code_searcher_class():
         str
             File statistic of current obj
         """
-        str_stats = ""
-        dict_num_of_code_lines_by_ext = \
-            self.get_dict_num_of_code_lines_by_extension()
-        str_stats += (
-            "Code lines: " +
-            str(dict_num_of_code_lines_by_ext["overall"]) +
-            "\n\n"
-        )
+        str_stats = "\n"
         str_stats += "Files Statistic of current code library:\n"
         # Print file statistic for every folder
-        for str_folder in self.dict_dict_str_file_by_filename_by_folder:
-            counter_int_files_by_extension = Counter()
-            dict_str_file_by_filename = \
-                self.dict_dict_str_file_by_filename_by_folder[str_folder]
-            list_filenames = list(dict_str_file_by_filename.keys())
-            str_stats += "--> For folder: " + str(str_folder)
-            str_stats += " found files with extensions:\n"
-            #####
+        for str_dir in self.dict_str_file_by_path_by_ext_by_dir:
+            str_stats_dir = ""
+            str_stats_dir += "--> For folder: " + str(str_dir) + "\n"
+            int_files_in_folder = 0
+            int_lines_in_folder = 0
+            str_stats_dir += "--> Files_found = {int_files_in_dir}  "
+            str_stats_dir += "Code_lines = {int_lines_in_dir}\n"
+            dict_str_file_by_path_by_ext = \
+                self.dict_str_file_by_path_by_ext_by_dir[str_dir]
             # Count number of files with every extension
-            for str_ext in self.list_str_file_extensions:
-                for str_filename in list_filenames:
-                    if str_filename.endswith(str_ext):
-                        counter_int_files_by_extension[str_ext] += 1
-                        # counter_int_files_by_extension["overall"] += 1
-            #####
-            # Add number of files with every extension to stats
-            for str_ext in counter_int_files_by_extension:
-                str_stats += (
-                    "----> {extension}: "
-                    "files_found = {int_files}  "
-                    "Code_lines = {int_lines}  "
+            for str_ext in dict_str_file_by_path_by_ext:
+                dict_str_file_by_path = dict_str_file_by_path_by_ext[str_ext]
+                int_code_lines = 0
+                for str_file_path in dict_str_file_by_path:
+                    str_full_file = dict_str_file_by_path[str_file_path]
+                    int_code_lines += \
+                        get_number_of_lines_in_string(str_full_file)
+
+                int_files = len(dict_str_file_by_path)
+                #####
+                int_files_in_folder += int_files
+                int_lines_in_folder += int_code_lines
+                #####
+                # Add number of files with every extension to stats
+                str_stats_dir += (
+                    "----> {extension}:  "
+                    "Files_found = {int_files};  "
+                    "Code_lines = {int_lines};  "
                     "\n"
                 ).format(
                     extension=str_ext,
-                    int_files=counter_int_files_by_extension[str_ext],
-                    int_lines=dict_num_of_code_lines_by_ext[str_ext],
+                    int_files=int_files,
+                    int_lines=int_code_lines,
                 )
+            str_stats += str_stats_dir.format(
+                int_files_in_dir=int_files_in_folder,
+                int_lines_in_dir=int_lines_in_folder,
+            )
             #####
             str_stats += "=" * 79 + "\n"
         if bool_is_to_print:
@@ -189,39 +244,104 @@ class code_searcher_class():
         return str_stats
 
     @check_type_of_arguments
-    def download_files(self):
-        """Download code files in dict_dict_str_file_by_filename_by_folder"""
-        float_time_start = time.time()
-        int_files_downloaded = 0
-        dict_dict_str_file_by_filename_by_folder = OrderedDict()
-        # For every folder where to look download code files
-        for str_folder_path in self.list_str_folders_where_to_look:
-            dict_str_full_file_by_rel_path = OrderedDict()
+    def get_dict_list_file_paths_by_ext_by_dir(self):
+        """Getting dict with pathes to asked files by ext by dir
+
+        Parameters
+        ----------
+
+        Returns
+        -------
+        dict
+            {"folder_path_1": {"file_extension_1": [file_path_1, ..], ..}, ..}
+        """
+        dict_list_file_paths_by_ext_by_dir = OrderedDict()
+        for str_dir_path in self.list_str_dirs_where_to_look:
+            dict_list_file_paths_by_ext = defaultdict(list)
             for str_ext in self.list_str_file_extensions:
-                dict_str_full_file_by_rel_path_tmp = \
-                    get_dict_str_full_file_by_rel_path(
-                        str_folder_path,
-                        str_extension_to_look_for=str_ext,
+                dict_list_file_paths_by_ext[str_ext] = \
+                    get_list_str_path_all_files_with_given_extension(
+                        str_dir_path,
+                        str_extension_to_look_for=str_ext
                     )
-                dict_str_full_file_by_rel_path.update(
-                    dict_str_full_file_by_rel_path_tmp
-                )
-                int_files_downloaded += len(dict_str_full_file_by_rel_path_tmp)
-            dict_dict_str_file_by_filename_by_folder[str_folder_path] = \
-                dict_str_full_file_by_rel_path
-        self.dict_dict_str_file_by_filename_by_folder = \
-            dict_dict_str_file_by_filename_by_folder
-        float_seconds_spent = round(time.time() - float_time_start, 2)
-        print(
-            "{} Files were downloaded in {} seconds".format(
-                int_files_downloaded,
-                float_seconds_spent
-            )
-        )
+            dict_list_file_paths_by_ext_by_dir[str_dir_path] = \
+                dict_list_file_paths_by_ext
+        return dict_list_file_paths_by_ext_by_dir
+
+    @check_type_of_arguments
+    def download_files(self):
+        """
+        For all files defined in self.dict_list_file_paths_by_ext_by_dir
+        download up to date versions of files (Efficient realization)
+
+        Parameters
+        ----------
+
+        Returns
+        -------
+        int
+            Number of files for which were downloaded new versions
+        """
+        int_new_files_downloaded = 0
+        #####
+        # For every folder where to look download code files
+        # Only if they were updated
+        for str_dir_path in self.dict_list_file_paths_by_ext_by_dir:
+            dict_list_file_paths_by_ext = \
+                self.dict_list_file_paths_by_ext_by_dir[str_dir_path]
+            #####
+            # Necessary for correct redownloading of library
+            if str_dir_path in self.dict_str_file_by_path_by_ext_by_dir:
+                dict_str_file_by_path_by_ext = \
+                    self.dict_str_file_by_path_by_ext_by_dir[str_dir_path]
+            else:
+                dict_str_file_by_path_by_ext = OrderedDict()
+            #####
+            for str_ext in dict_list_file_paths_by_ext:
+                list_str_file_paths = dict_list_file_paths_by_ext[str_ext]
+                #####
+                # Necessary for correct redownloading of library
+                if str_ext in dict_str_file_by_path_by_ext:
+                    dict_str_file_by_path = \
+                        dict_str_file_by_path_by_ext[str_ext]
+                else:
+                    dict_str_file_by_path = OrderedDict()
+                #####
+                for str_f_path in list_str_file_paths:
+                    # Check if file was deleted
+                    # then delete it from dict_str_file_by_path
+                    if not os.path.exists(str_f_path):
+                        dict_str_file_by_path[str_f_path] = ""
+                        continue
+                    #####
+                    # Check if file was modified and if so redownload it
+                    float_time_file_changed = os.stat(str_f_path).st_mtime
+                    float_time_file_mod_before = \
+                        self.dict_time_file_changed_by_path[str_f_path]
+                    if float_time_file_mod_before != float_time_file_changed:
+                        dict_str_file_by_path[str_f_path] = \
+                            get_file_as_string(str_f_path)
+                        self.dict_time_file_changed_by_path[str_f_path] = \
+                            float_time_file_changed
+                        int_new_files_downloaded += 1
+                dict_str_file_by_path_by_ext[str_ext] = dict_str_file_by_path
+            #####
+            self.dict_str_file_by_path_by_ext_by_dir[str_dir_path] = \
+                dict_str_file_by_path_by_ext
+        return int_new_files_downloaded
 
     @check_type_of_arguments
     def update_files(self):
-        """Redownload code files in dict_dict_str_file_by_filename_by_folder"""
+        """Re-search for all files in the folders and download them
+
+        Parameters
+        ----------
+        Returns
+        -------
+        """
+        print("Updating all files (it can be a long process, please wait.)")
+        self.dict_list_file_paths_by_ext_by_dir = \
+            self.get_dict_list_file_paths_by_ext_by_dir()
         self.download_files()
 
     @check_type_of_arguments
@@ -231,7 +351,7 @@ class code_searcher_class():
             bool_is_to_search_case_sensitive=True,
     ):
         """Searching some code inside whole library and
-        print all occurences in formatted way
+        print all occurences in nice formatted way
 
         Parameters
         ----------
@@ -245,44 +365,68 @@ class code_searcher_class():
         int
             times occurences of code found in whole library
         """
+        # 0) Reload all modified files, to have right versions
+        self.download_files()
+        # 1) If not neccesary to search case sensitive, then lower everything
         if not bool_is_to_search_case_sensitive:
             str_code_to_search = str_code_to_search.lower()
         int_occurrences_found = 0
+        print("=" * 79)
         # For every folder searching through all files inside folder
-        for str_folder_name in self.dict_dict_str_file_by_filename_by_folder:
-            dict_str_full_file_by_rel_path = \
-                self.dict_dict_str_file_by_filename_by_folder[str_folder_name]
-            print("For folder: {folder}\nFiles found: {int_files}".format(
-                folder=str_folder_name,
-                int_files=len(dict_str_full_file_by_rel_path)),
-            )
-            # For every file search occurences of asked code
-            for str_filename in dict_str_full_file_by_rel_path:
-                str_full_file = dict_str_full_file_by_rel_path[str_filename]
-                if not bool_is_to_search_case_sensitive:
-                    str_full_file = str_full_file.lower()
-                list_str_file_splitted = enumerate(str_full_file.splitlines())
-                # Line by line searching for asked code
-                for int_line_num, str_line in list_str_file_splitted:
-                    if str_code_to_search in str_line:
-                        print(
-                            "--> ", int_occurrences_found,
-                            ") Found in: ", str_filename
-                        )
-                        print(
-                            "----> line: ", int_line_num,
-                            " Code_line: ", str_line.strip()
-                        )
-                        int_occurrences_found += 1
+        for str_dir in self.dict_str_file_by_path_by_ext_by_dir:
+            dict_str_file_by_path_by_ext = \
+                self.dict_str_file_by_path_by_ext_by_dir[str_dir]
+            print("For folder: {folder}".format(folder=str_dir))
+            for str_ext in dict_str_file_by_path_by_ext:
+                if len(self.list_str_file_extensions) > 1:
+                    print("")
+                print("--> For extension: {extension}".format(
+                    extension=str_ext
+                ))
+                bool_is_entry_found_for_cur_ext = False
+                dict_str_file_by_path = dict_str_file_by_path_by_ext[str_ext]
+                # For every file search occurences of asked code
+                for str_file_path in dict_str_file_by_path:
+                    str_rel_path = os.path.relpath(str_file_path, str_dir)
+                    str_full_file = dict_str_file_by_path[str_file_path]
+                    if not bool_is_to_search_case_sensitive:
+                        str_full_file = str_full_file.lower()
+                    list_str_file_splitted = enumerate(
+                        str_full_file.splitlines()
+                    )
+                    #####
+                    # Line by line searching for asked code
+                    bool_is_entry_found_for_cur_file = False
+                    for int_line_num, str_line in list_str_file_splitted:
+                        if str_code_to_search in str_line:
+
+                            if not bool_is_entry_found_for_cur_file:
+                                bool_is_entry_found_for_cur_file = True
+                                bool_is_entry_found_for_cur_ext = True
+                                print("----> Found in: ", str_rel_path)
+                            print(
+                                "------> ", int_occurrences_found,
+                                ") line:", int_line_num,
+                                " Code_line:", str_line.strip()
+                            )
+                            int_occurrences_found += 1
+                #####
+                if not bool_is_entry_found_for_cur_ext:
+                    print("----> NOTHING FOUND.")
+            if len(self.list_str_file_extensions) > 1:
+                print("")
             #####
             if int_occurrences_found:
-                print("#" * 79)
+                print("=" * 79)
         print("Overall occurrences found: ", int_occurrences_found)
         return int_occurrences_found
 
     @check_type_of_arguments
-    def get_dict_times_functions_used(self):
+    def get_dict_times_py_functions_used(self):
         """Getting {function_name: times_function_used, ...}
+
+        Parameters
+        ----------
 
         Returns
         -------
@@ -290,18 +434,20 @@ class code_searcher_class():
             {function_name: times_function_used, ...}
         """
         defdict_times_functions_used = defaultdict(int)
-        set_str_funcs_defined_in_the_library = \
+        set_str_all_funcs_defined = \
             self.get_names_of_all_functions_defined_in_the_py_library()
         # For every folder searching through all files inside folder
-        for str_folder in self.dict_dict_str_file_by_filename_by_folder:
-            dict_str_full_file_by_rel_path = \
-                self.dict_dict_str_file_by_filename_by_folder[str_folder]
-            # For every file searching all occurences of function definition
-            for str_filename in dict_str_full_file_by_rel_path:
-                str_full_file = dict_str_full_file_by_rel_path[str_filename]
-                for str_func_to_count in set_str_funcs_defined_in_the_library:
-                    defdict_times_functions_used[str_func_to_count] += \
-                        str_full_file.count(str_func_to_count + "(")
+        for str_dir in self.dict_str_file_by_path_by_ext_by_dir:
+            dict_str_file_by_path_by_ext = \
+                self.dict_str_file_by_path_by_ext_by_dir[str_dir]
+            for str_ext in dict_str_file_by_path_by_ext:
+                dict_str_file_by_path = dict_str_file_by_path_by_ext[str_ext]
+                # For every file searching all occurences of function definition
+                for str_filename in dict_str_file_by_path:
+                    str_full_file = dict_str_file_by_path[str_filename]
+                    for str_func_to_count in set_str_all_funcs_defined:
+                        defdict_times_functions_used[str_func_to_count] += \
+                            str_full_file.count(str_func_to_count + "(")
         #####
         # Subtract 1 from times every function used, as its function definition
         for str_func_name in defdict_times_functions_used:
@@ -309,15 +455,18 @@ class code_searcher_class():
         return defdict_times_functions_used
 
     @check_type_of_arguments
-    def get_names_of_all_functions_defined_but_never_used(self):
+    def get_names_of_all_py_functions_defined_but_never_used(self):
         """Getting list of functions defined inside code but never used
+
+        Parameters
+        ----------
 
         Returns
         -------
         list
             list names of never used functions inside the code library
         """
-        defdict_times_functions_used = self.get_dict_times_functions_used()
+        defdict_times_functions_used = self.get_dict_times_py_functions_used()
         list_str_never_used_functions = []
         for str_func_name in defdict_times_functions_used:
             if not defdict_times_functions_used[str_func_name]:
@@ -329,31 +478,6 @@ class code_searcher_class():
         return sorted(list_str_never_used_functions)
 
     @check_type_of_arguments
-    def get_dict_num_of_code_lines_by_extension(self):
-        """Getting dict number of not empty lines by file extension type
-
-        Returns
-        -------
-        dict
-            {".py": X, ".ipynb": Y, ...}
-        """
-        counter_num_of_code_lines_by_ext = Counter()
-        counter_num_of_code_lines_by_ext["overall"] = 0
-        for str_folder in self.dict_dict_str_file_by_filename_by_folder:
-            dict_str_file_by_filename = \
-                self.dict_dict_str_file_by_filename_by_folder[str_folder]
-            for str_filename in dict_str_file_by_filename:
-                str_ext = "." + os.path.basename(str_filename).split(".")[-1]
-                str_whole_file = dict_str_file_by_filename[str_filename]
-                int_lines_found = \
-                    str_whole_file.count('\n') + 1
-                counter_num_of_code_lines_by_ext[str_ext] += \
-                    int_lines_found
-                counter_num_of_code_lines_by_ext["overall"] += \
-                    int_lines_found
-        return counter_num_of_code_lines_by_ext
-
-    @check_type_of_arguments
     def get_number_of_lines_in_the_library(self):
         """Getting number of not empty lines in whole library
 
@@ -363,18 +487,23 @@ class code_searcher_class():
             Number of Not empty code lines in the library
         """
         int_lines_of_code_already_found = 0
-        for str_folder in self.dict_dict_str_file_by_filename_by_folder:
-            dict_str_file_by_filename = \
-                self.dict_dict_str_file_by_filename_by_folder[str_folder]
-            for str_filename in dict_str_file_by_filename:
-                str_whole_file = dict_str_file_by_filename[str_filename]
-                int_lines_of_code_already_found += \
-                    str_whole_file.count('\n') + 1
+        for str_dir in self.dict_str_file_by_path_by_ext_by_dir:
+            dict_str_file_by_path_by_ext = \
+                self.dict_str_file_by_path_by_ext_by_dir[str_dir]
+            for str_ext in dict_str_file_by_path_by_ext:
+                dict_str_file_by_path = dict_str_file_by_path_by_ext[str_ext]
+                for str_file_path in dict_str_file_by_path:
+                    str_whole_file = dict_str_file_by_path[str_file_path]
+                    int_lines_of_code_already_found += \
+                        get_number_of_lines_in_string(str_whole_file)
         return int_lines_of_code_already_found
 
     @check_type_of_arguments
     def get_names_of_all_functions_defined_in_the_py_library(self):
         """Getting set names of all functions defined in the library
+
+        Parameters
+        ----------
 
         Returns
         -------
@@ -382,15 +511,17 @@ class code_searcher_class():
             set names of all functions defined in the library
         """
         list_str_all_functions_defined_in_the_library = []
-        for str_folder in self.dict_dict_str_file_by_filename_by_folder:
-            dict_str_full_file_by_rel_path = \
-                self.dict_dict_str_file_by_filename_by_folder[str_folder]
-            for str_filename in dict_str_full_file_by_rel_path:
-                str_full_file = dict_str_full_file_by_rel_path[str_filename]
-                list_str_all_functions_defined_in_the_library += \
-                    get_names_of_all_functions_defined_in_py_code(
-                        str_full_file
-                    )
+        for str_dir in self.dict_str_file_by_path_by_ext_by_dir:
+            dict_str_file_by_path_by_ext = \
+                self.dict_str_file_by_path_by_ext_by_dir[str_dir]
+            for str_ext in dict_str_file_by_path_by_ext:
+                dict_str_file_by_path = dict_str_file_by_path_by_ext[str_ext]
+                for str_file_path in dict_str_file_by_path:
+                    str_full_file = dict_str_file_by_path[str_file_path]
+                    list_str_all_functions_defined_in_the_library += \
+                        get_names_of_all_functions_defined_in_py_code(
+                            str_full_file
+                        )
         set_str_all_functions_defined_in_the_library = \
             set(list_str_all_functions_defined_in_the_library)
         print(
@@ -403,51 +534,60 @@ class code_searcher_class():
     def get_set_str_names_of_all_py_files(self):
         """Getting set names of all .py files inside the library
 
+        Parameters
+        ----------
+
         Returns
         -------
         set
             set names of all .py files inside the library
         """
         list_names_of_all_py_files_in_library = []
-        for str_folder_where_to_look in self.list_str_folders_where_to_look:
+        for str_dir_where_to_look in self.list_str_dirs_where_to_look:
             list_names_of_all_py_files_in_library += \
                 get_list_str_filenames_of_all_files_with_given_extension(
-                    str_folder_where_to_look,
-                    str_ext_to_look_for=".py"
+                    str_dir_where_to_look,
+                    str_extension_to_look_for=".py"
                 )
         return set(list_names_of_all_py_files_in_library)
 
     @check_type_of_arguments
-    def get_all_imported_py_modules_in_the_library(self):
+    def get_list_all_imported_py_modules_in_the_library(self):
         """Getting list of all modules imported in the library
+
+        Parameters
+        ----------
+
         Returns
         -------
         list
             list names of all imported modules found
         """
         list_imported_modules_found = []
-        for str_folder in self.dict_dict_str_file_by_filename_by_folder:
-            dict_my_python_file_by_file_name = \
-                self.dict_dict_str_file_by_filename_by_folder[str_folder]
-            print(
-                "Overall files found: ",
-                len(dict_my_python_file_by_file_name)
-            )
-            #####
-            # One by one dealing with every file
-            for str_filename in dict_my_python_file_by_file_name:
-                str_full_code_of_one_py_file = \
-                    dict_my_python_file_by_file_name[str_filename]
-                list_imported_modules_found += \
-                    get_list_modules_imported_in_py_code(
-                        str_full_code_of_one_py_file
-                    )
+        for str_dir in self.dict_str_file_by_path_by_ext_by_dir:
+            dict_str_file_by_path_by_ext = \
+                self.dict_str_file_by_path_by_ext_by_dir[str_dir]
+            for str_ext in dict_str_file_by_path_by_ext:
+                dict_str_file_by_path = dict_str_file_by_path_by_ext[str_ext]
+                #####
+                # One by one dealing with every file
+                for str_filename in dict_str_file_by_path:
+                    str_full_code_of_one_py_file = \
+                        dict_str_file_by_path[str_filename]
+                    list_imported_modules_found += \
+                        get_list_modules_imported_in_py_code(
+                            str_full_code_of_one_py_file
+                        )
         list_imported_modules_found = list(set(list_imported_modules_found))
         return list_imported_modules_found
 
     @check_type_of_arguments
     def get_list_of_all_outer_modules_used_in_the_library(self):
         """Getting list of all OUTER modules imported in the library
+
+        Parameters
+        ----------
+
         Returns
         -------
         list
@@ -457,7 +597,7 @@ class code_searcher_class():
         set_str_names_of_all_py_files = \
             self.get_set_str_names_of_all_py_files()
         list_imported_modules_found = \
-            self.get_all_imported_py_modules_in_the_library()
+            self.get_list_all_imported_py_modules_in_the_library()
         print(
             "Overall unique modules imported: ",
             len(list_imported_modules_found)