+++ CHANGELOG.rst
@@ -2,8 +2,14 @@
 Changelog
 =========
 
-Version 0.1.1 - 0.1.N
+Version 0.1.6 (Huge changes)
 =================================
+- New way of dealing with files, now instead of dowloading all files again and again
+  only changes from the initial download will be download (git style)
+- New much more pretty format of output for function: search_all_occurrences_of_the_code_in_the_library
+
+Version 0.1.1 - 0.1.5 (Bug fixes)
+===================================
 - FIX: Updated docs
 - FIX: minor bags were fixed.